Extends documentation for include/exclude scopes
The problem was that the property names are in plural (speaking from
included/ecluded scopes). So as a user I assume that I can add multiple
scopes, but nowhere in the documentation the syntax for that is mentioned.

After digging through the source code I found that the scopes are delimitted
by comma.

So I extend the documentation for the properties

- license.excludedScopes
- license.includedScopes

how to add multiple scopes.