Use 'GAV' for result keys in LicensedArtifactResolver
Fixes #413

In the result mapping we need to use "GAV" to be compatible with
the other classes in the plugin. While the exclude/includeArtifacts
should still use the Artifact#getId, since the getDependencyTrail
also uses that format, which is considered when we exclude all the
transitive deps of excluded artifacts.

Note that the added it is based on the already existing
download-licenses-include-exclude-types, but with a small change
in the settings to use excludeTransitiveDependencies instead.